warehouse/postgresql/requirements.txt
boto3
pandas
pyarrow
psycopg2-binary
s3fs